SEQNG-883 Fixed issue qith resuming queued sequence, added tests.
