+++ modules/acm/src/test/java/edu/gemini/epics/acm/test/CaStatusAcceptorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2018 Association of Universities for Research in Astronomy, Inc. (AURA)
+ * Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
  * For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
  */
