+++ modules/acm/src/test/scala/edu/gemini/epics/acm/ObserveStateSpec.scala
@@ -1,4 +1,4 @@
-// Copyright (c) 2016-2018 Association of Universities for Research in Astronomy, Inc. (AURA)
+// Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
 package edu.gemini.epics.acm
